feat(dmm-tv): party button on playback page + same-tab create nav
- 再生ページ（/vod/playback）の お気に入り/シェア アクション行にも「パーティー作成」
  アイコンを追加（通常視聴 normal モードのみ）。クリックでその場で作成サイドバーを開く
  （setMode("create") は visible も true にするため遷移・リロード不要）。
- アクション行への注入処理を共有 dom/dmmPartyButton.ts に切り出し、詳細ページと再生
  ページで共用（お気に入り/シェアの行を aria-label から辿って冪等注入・SPA 追従）。
- 詳細ページの「パーティー作成」を別タブ（window.open）から同じタブ遷移
  （window.location.href）へ戻す。
